Ot138 24 form login (#3)
* create styled component authHeader

create header component to be used on login and sign up form views. The image is temporary. The styles needs to be unified with the rest of the api

* create styled login view

create login view, with temporary styles that needs to be unified and refactored with the rest of the app. Includes validation, but not the actual session login