Ot138-20-setup-alerts (#11)
* OT138-31-http-services (#2)

* Add request interceptor to axios instance to update Authorization header

* Refactor api request interceptor code

* Add console warnings in case env variables are not defined

* Ot138 24 form login (#3)

* create styled component authHeader

create header component to be used on login and sign up form views. The image is temporary. The styles needs to be unified with the rest of the api

* create styled login view

create login view, with temporary styles that needs to be unified and refactored with the rest of the app. Includes validation, but not the actual session login

* Checking if ci/cd is working

* Ot138 23 signup form (#6)

* add styled signup form

add signup view with form, inline-styles (that has to be refactored), and validation with the formik library

* refactor styles from login and signup views. move styles to external script

* fix header style

* Adding login and register render route

* Minor fix

* Rename authHeader.js to AuthHeader.js

* fix incorrect link to Signup view (#8)

* Adding Loader Component (#10)

* Minor Fixs

* Update Readme

* Setup Alerts Component

Co-authored-by: Facundo Juarez <31292867+facundojuarez2k@users.noreply.github.com>
Co-authored-by: Nicolás Oleinizak <78864158+nicolasoleinizak@users.noreply.github.com>